Rename Instruction to Procedure
