+++ .idea/runConfigurations/javafx_Main.xml
@@ -1,7 +1,7 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="javafx_Main" type="Application" factoryName="Application">
     <option name="MAIN_CLASS_NAME" value="javafx.ApplicationStarter" />
-    <module name="spio2023.calibration.view_javafx" />
+    <module name="view-javafx" />
     <method v="2">
       <option name="Make" enabled="true" />
     </method>
