+++ model/src/main/java/procedure/Procedure.java
@@ -1,0 +1,10 @@
+package procedure;
+
+import device.Device;
+import procedure.step.Step;
+
+import java.util.List;
+
+public record Procedure(Device referenceDevice, Device checkedDevice, List<Double> controlPoints, List<Step> steps) {
+
+}
