+++ model/src/main/java/procedure/Settings.java
@@ -1,4 +1,4 @@
-package instruction;
+package procedure;
 
 public record Settings(int measurementsNumber) {
 }
