+++ model/src/main/java/procedure/step/CalculateResultsStep.java
@@ -1,4 +1,4 @@
-package instruction.step;
+package procedure.step;
 
 import scope.AccuracyScope;
