+++ model/src/main/java/procedure/step/Results.java
@@ -1,4 +1,4 @@
-package instruction.step;
+package procedure.step;
 
 import java.util.ArrayList;
 import java.util.Collections;
